Refactoring the testing example unit tests
